fix loading bug

+++ mocap/datasets/h36m.py
@@ -58,7 +58,7 @@ def get3d_fixed(actor, action, sid):
     H36M_DA.acquire_fixed_skeleton()
     fname_binary = join(join(data_dir, 'fixed_skeleton'), actor + '_' + action + '_' + str(sid) + '.npy')
     if isfile(fname_binary):
-        seq = np.load(fname_binary, dtype=np.float32)
+        seq = np.load(fname_binary)
     else:
         fname = join(join(data_dir, 'fixed_skeleton'), actor + '_' + action + '_' + str(sid) + '.txt')
         seq = np.loadtxt(fname, dtype=np.float32)
